refactor: the folder structure
