+++ src/invoices/VersionHistory/InvoiceVersionHistory/VersionHistory.js
@@ -13,16 +13,16 @@ import {
   VersionViewContextProvider,
 } from '@folio/stripes-acq-components';
 
-import { INVOICE_ROUTE } from '../../common/constants';
+import { INVOICE_ROUTE } from '../../../common/constants';
 import {
   useInvoice,
   useInvoiceVersions,
   useSelectedInvoiceVersion,
-} from '../../common/hooks';
+} from '../../../common/hooks';
 import {
   HIDDEN_INVOICE_FIELDS,
   INVOICE_FIELDS_LABEL_MAP,
-} from './constants';
+} from '../constants';
 import { VersionHistoryView } from './VersionHistoryView';
 
 const VersionHistory = ({
