+++ src/invoices/VersionHistory/InvoiceVersionHistory/index.js
@@ -1,0 +1,1 @@
+export { default as InvoiceVersionHistory } from './VersionHistory';
