--- src/invoices/VersionHistory/VersionHistoryView/VersionHistoryViewAdjustments/index.js
@@ -1,1 +1,0 @@
-export { VersionHistoryViewAdjustments } from './VersionHistoryViewAdjustments';
