+++ src/invoices/VersionHistory/components/VersionHistoryAdjustments/index.js
@@ -1,0 +1,1 @@
+export { VersionHistoryAdjustments } from './VersionHistoryAdjustments';
