+++ src/invoices/VersionHistory/components/index.js
@@ -1,0 +1,1 @@
+export { VersionHistoryAdjustments } from './VersionHistoryAdjustments';
