+++ src/invoices/VersionHistory/index.js
@@ -1,1 +1,2 @@
-export { default as VersionHistory } from './VersionHistory';
+export { InvoiceVersionHistory } from './InvoiceVersionHistory';
+export { InvoiceLineVersionHistory } from './InvoiceLineVersionHistory';
